requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv
pandas
